added login form to component
